+++ changelog.md
@@ -1,5 +1,6 @@
 ### Next Version
 * Upgraded Kotlin from 1.7.10 to 1.7.20
+* Upgraded cloudevents from 2.3.0 to 2.4.0
 
 ### 0.14.8 (2022-10-10)
 * Fixed NPE issue in the toString() method in `org.occurrent.eventstore.api.StreamVersionWriteCondition` when stream condition was `any`.