+++ changelog.md
@@ -2,6 +2,7 @@
 * Upgraded Kotlin from 1.7.10 to 1.7.20
 * Upgraded cloudevents from 2.3.0 to 2.4.0
 * Upgraded Spring Boot from 2.7.3 to 2.7.4
+* Changed toString() on StreamVersionWriteCondition when condition is null from "any stream version" to "any"
 
 ### 0.14.8 (2022-10-10)
 * Fixed NPE issue in the toString() method in `org.occurrent.eventstore.api.StreamVersionWriteCondition` when stream condition was `any`.